xlunch: genentries.desktop.sh: Execute genentries under bash
There is an error in the upstream genentries.desktop.sh script,
in which they invoke genentries under `sh` — which in Ubuntu is dash.
genentries is a bash script, not a posix shell script.

Take genentries.desktop.sh under control here and modify it to work,
until such time as the upstream problem is fixed.